neural_network_mine_detection/requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
Pillow>=8.0.0
matplotlib>=3.3.2
scikit-learn>=0.23.2
opencv-python>=4.4.0 